Move record-keeping classes to their own directory
