+++ src/game/records/__init__.py
@@ -1,0 +1,3 @@
+from .game_record import Game_Record
+from .players import Players
+from .triviaset import Trivia_Set
